+++ tests/Containers/DataBagTest.php
@@ -115,6 +115,29 @@ class DataBagTest extends PHPUnit_Framework_TestCase
     }
 
     /**
+     * @coversNothing
+     */
+    public function testCanUnsetProperty()
+    {
+        // ----------------------------------------------------------------
+        // setup your test
+
+        $obj = new DataBag;
+        $obj->test1 = 1;
+        $this->assertTrue(isset($obj->test1));
+
+        // ----------------------------------------------------------------
+        // perform the change
+
+        unset($obj->test1);
+
+        // ----------------------------------------------------------------
+        // test the results
+
+        $this->assertFalse(isset($obj->test1));
+    }
+
+    /**
      * @covers ::__get
      */
     public function testCanUseDotNotationToGetProperties()
@@ -186,6 +209,33 @@ class DataBagTest extends PHPUnit_Framework_TestCase
         // test the results
 
         $this->assertTrue(isset($obj->{'child1.child2'}));
+    }
+
+    /**
+     * @covers ::__unset
+     */
+    public function testCanUnsetPropertyUsingDotNotationSupport()
+    {
+        // ----------------------------------------------------------------
+        // setup your test
+
+        $obj = new DataBag;
+        $obj->child1 = new DataBag;
+        $obj->child1->child2 = 100;
+
+        $this->assertTrue(isset($obj->child1->child2));
+        $this->assertTrue(isset($obj->{'child1.child2'}));
+
+        // ----------------------------------------------------------------
+        // perform the change
+
+        unset($obj->{'child1.child2'});
+
+        // ----------------------------------------------------------------
+        // test the results
+
+        $this->assertTrue(isset($obj->child1));
+        $this->assertFalse(isset($obj->child1->child2));
     }
 
     /**